[Gold II] Title: 피보나치 수 6, Time: 0 ms, Memory: 1180 KB -BaekjoonHub

+++ 백준/Gold/11444. 피보나치 수 6/피보나치 수 6.cc
@@ -3,11 +3,10 @@
 
 using namespace std;
 
-
 long long n;
-map <long long, long long> m;
+map <long long, long> m;
 
-long long fibo(long long a) {
+long fibo(long long a) {
 	if (m[a]) {
 		return m[a];
 	}
